+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Feature
 - Added support for multiple template engines (Twig & Blade)
 - Each pattern folder can also include an optional `init.php` that gets loaded before any data making it easier to scope view composers
+- Patterns now include raw template rendering
+- Patterns now show the data needed to render them
 
 ### Breaking
 - New Handlebars engine drops support for `.handlebars` extensions, only `.hbs` now supported out of the box